catalog_service/requirements.txt
django>=5.2
djangorestframework>=3.15
gunicorn>=21.2
python-jose[cryptography]>=3.3.0
requests>=2.31